pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ynab-mcp"
version = "0.1.0"
description = "A Model Context Protocol server for interacting with YNAB"
readme = "README.md"
requires-python = ">=3.13"
license = {text = "MIT"}
dependencies = [
    "fastmcp>=0.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "httpx>=0.26.0",
]

[project.scripts]
ynab-mcp = "ynab_mcp.server:main"

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
src_paths = ["src", "tests"]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "respx>=0.22.0",
]
